requirements.txt
attrs==24.3.0
blinker==1.8.2
click==8.1.8
dnspython==2.6.1
flasgger==0.9.7.1
flask==3.0.3
importlib-metadata==8.5.0
importlib-resources==6.4.5
itsdangerous==2.2.0
jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
mistune==3.1.0
packaging==24.2
pkgutil-resolve-name==1.3.10
pymongo==4.10.1
python-dotenv==1.0.1
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.20.1
six==1.17.0
typing-extensions==4.12
werkzeug==3.0.6
zipp==3.20.2
